requirements.txt
Adafruit-Blinka==6.3.2
adafruit-circuitpython-bme280==2.5.4
adafruit-circuitpython-busdevice==5.0.6
Adafruit-PlatformDetect==3.3.0
Adafruit-PureIO==1.1.8
pkg-resources==0.0.0
pyftdi==0.52.9
pyserial==3.5
pyusb==1.1.1
rpi-ws281x==4.2.5
RPi.GPIO==0.7.0
sysv-ipc==1.1.0

python-crontab~=2.5.1
rpi~=0.0.1